Avion/requirements-dev.txt
-r requirements.txt

# Additional Development Dependencies
pytest-watch>=4.2.0
pytest-sugar>=0.9.4
ipython>=8.0.0
jupyter>=1.0.0
pre-commit>=2.17.0 